fix openapi
